split cellxgene_census-dependent code out of utils.py into census_utils.py
utils.py no longer imports cellxgene_census, so it can run in the
scanpyenv-minimal container which lacks that dependency. setup.py and
get_census_adata.py now pull from census_utils; process_query.py and
scvi_classify.py drop their now-redundant cellxgene_census imports.
SETUP_SCVI points at the census-minimal image accordingly. setup()
also switches from wget to urllib with an unverified SSL context.
