+++ bin/setup.py
@@ -1,46 +1,15 @@
-#!/user/bin/python3
+#!/usr/bin/env python3
 
-
-
-from pathlib import Path
-import scanpy as sc
-import numpy as np
-import pandas as pd
-import anndata as ad
-import cellxgene_census
-import scvi
-from scipy.sparse import csr_matrix
-import cellxgene_census
-import cellxgene_census.experimental
-import scvi
-from sklearn.ensemble import RandomForestClassifier
-from utils import *
-from pathlib import Path
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.model_selection import train_test_split
-from sklearn.metrics import accuracy_score, classification_report, confusion_matrix, roc_curve, auc
-from sklearn.preprocessing import label_binarize
-import yaml
-import matplotlib.pyplot as plt
-import seaborn as sns
 import argparse
-import os
+from census_utils import setup
 
 
-
-# Function to parse command line arguments
 def parse_arguments():
-    parser = argparse.ArgumentParser(description="Download model file based on organism, census version, and tree file.")
-    parser.add_argument('--organism', type=str, default='homo_sapiens', help='Organism name (e.g., homo_sapiens)')
-    parser.add_argument('--census_version', type=str, default='2024-07-01', help='Census version (e.g., 2024-07-01)')
+    parser = argparse.ArgumentParser(description="Download scVI model for a given organism and census version.")
+    parser.add_argument('--organism', type=str, default='homo_sapiens')
+    parser.add_argument('--census_version', type=str, default='2024-07-01')
     return parser.parse_args()
 
-# Parse command line arguments
+
 args = parse_arguments()
-
-# Set organism and census_version from arguments
-organism = args.organism
-census_version = args.census_version
-
-# Get model file link and download
-model_path = setup(organism=organism, version=census_version)
+model_path = setup(organism=args.organism, version=args.census_version)
