+++ bin/utils.py
@@ -4,8 +4,6 @@ import pandas as pd
 import numpy as np
 import scanpy as sc
 import random
-import cellxgene_census
-import cellxgene_census.experimental
 import os
 import anndata as ad
 import scvi
@@ -26,123 +24,6 @@ import warnings
 warnings.filterwarnings("ignore")
 
 
-def setup(organism="homo_sapiens", version="2024-07-01"):
-    organism=organism.replace(" ", "_") 
-    #census = cellxgene_census.open_soma(census_version=version)
-    outdir = f"scvi-{organism}-{version}"  # Concatenate strings using f-string
-    
-    # Create the output directory if it doesn't exist
-    if not os.path.exists(outdir):
-        os.makedirs(outdir)
-
-    # Check if the model file exists
-    model_file_path = os.path.join(outdir, "model.pt")
-    #if not os.path.exists(model_file_path):
-        # Get scVI model info
-    scvi_info = cellxgene_census.experimental.get_embedding_metadata_by_name(
-            embedding_name="scvi",
-            organism=organism,
-            census_version=version,
-        )
-
-        # Extract the model link
-    model_link = scvi_info["model_link"]
-    date = model_link.split("/")[5]
-    url = os.path.join("https://cellxgene-contrib-public.s3.us-west-2.amazonaws.com/models/scvi/", date, organism, "model.pt")
-
-    # Download the model using wget if it doesn't already exist
-    subprocess.run(["wget", "--no-check-certificate", "-q", "-O", model_file_path, url])
-# else:
-     #   print(f"File already exists at {model_file_path}, skipping download.")
-
-    return(outdir)
-
-#clean up cellxgene ontologies
-
-def rename_cells(obs, rename_file="/space/grp/rschwartz/rschwartz/cell_annotation_cortex.nf/meta/rename_cells_mmus.tsv"):
-    # change to handle author_cell_type as first column name of rename df (distinguishes between author and cellxgene annotations)
-    rename_df = pd.read_csv(rename_file, sep=None)
-    rename_key = rename_df.columns[0]
-    rename_cell_type = rename_df.columns[1]
-    # Filter cells to keep only those with valid new cell types
-    # this replaces the "restricted cell types" command line argument
-    # not working for "glutamatergic neuron" in human
-    
-    obs = obs[obs[rename_key].isin(rename_df[rename_key])]
- 
-    # Create mapping dictionaries
-    rename_mapping = dict(zip(rename_df[rename_key], rename_df[rename_cell_type]))
-    ontology_mapping = dict(zip(rename_df[rename_cell_type], rename_df[rename_cell_type + "_uri"]))
-    
-    # Apply renaming
-    obs[rename_cell_type] = obs[rename_key].replace(rename_mapping)  
-    obs[rename_cell_type + "_uri"] = obs[rename_cell_type].map(ontology_mapping)
-
-    return obs
-
-
-# Subsample x cells from each cell type if there are n>x cells present
-# calls "rename cells" to filter out generic/ambiguous cell types
-# ensures equal representation of cell types in reference
-def subsample_cells(data, filtered_ids, subsample=500, seed=42, organism="Homo sapiens", 
-                    rename_file="/space/grp/rschwartz/rschwartz/cell_annotation_cortex.nf/meta/rename_cells_mmus.tsv"):
-    """
-    Subsample up to `subsample` cells per cell type, ensuring balanced representation.
-    Filter and rename cells before sub-sampling (see rename_cells).
-    """
-    random.seed(seed)         # For `random`
-    np.random.seed(seed)      # For `numpy`
-    scvi.settings.seed = seed # For `scvi`
-    
-    # Filter data based on filtered_ids
-    obs = data[data['soma_joinid'].isin(filtered_ids)]
-
-    obs = rename_cells(obs, rename_file=rename_file)
-    celltypes = obs["cell_type"].unique()
-    print(obs["cell_type"].value_counts().reset_index())
-    final_idx = []
-    for celltype in celltypes:
-        celltype_ids = obs[obs["cell_type"] == celltype]['soma_joinid'].values
-        # Sample if there are enough observations, otherwise take all
-        if len(celltype_ids) > subsample:
-            subsampled_cell_idx = random.sample(list(celltype_ids), subsample)
-        else:
-            subsampled_cell_idx = celltype_ids.tolist()
-        # Append subsampled indices to final list
-        final_idx.extend(subsampled_cell_idx)
-
-    # Return final indices
-    return final_idx
-
-
-def get_original_celltypes(columns_file="/space/grp/rschwartz/rschwartz/nextflow_eval_pipeline/meta/author_cell_annotations/original_celltype_columns.tsv", 
-                           author_annotations_path="/space/grp/rschwartz/rschwartz/nextflow_eval_pipeline/meta/author_cell_annotations"):
-    original_celltype_columns = pd.read_csv(columns_file, sep=None)
-
-    original_celltypes = {}
-    for file in os.listdir(author_annotations_path):
-        if "obs.tsv" in file:
-            dataset_title = file.split(".")[0]
-            og_obs = pd.read_csv(os.path.join(author_annotations_path, file), sep=None)
-            # check if all observation_joinid are unique
-            assert og_obs["observation_joinid"].nunique() == og_obs.shape[0]
-            og_column = original_celltype_columns[original_celltype_columns["dataset_title"] == dataset_title]["author_cell_type"].values[0]
-            og_obs["author_cell_type"] = og_obs[og_column]
-            original_celltypes[dataset_title] = og_obs
-
-    for dataset_title, obs in original_celltypes.items():
-        #original_celltypes[dataset_title]["new_dataset_title"] = dataset_title
-        original_celltypes[dataset_title]["new_observation_joinid"] = original_celltypes[dataset_title]["observation_joinid"].apply(lambda x: f"{dataset_title}_{x}")
-    
-        # concat all original_celltypes
-    aggregate_obs = pd.concat([original_celltypes[ref_name] for ref_name in original_celltypes.keys()])
-    # find duplicate observation_joinid in aggregate_obs
-    duplicate_observation_joinid = aggregate_obs[aggregate_obs["new_observation_joinid"].duplicated()]
-    duplicate_observation_joinid.columns
-    assert aggregate_obs["new_observation_joinid"].nunique() == aggregate_obs.shape[0]
-   
-    return aggregate_obs
-
 def relabel(adata, relabel_path, join_key="", sep=None):
     # Read the relabel table from the file
     relabel_df = pd.read_csv(relabel_path, sep=sep)  # Adjust the separator as needed
@@ -159,124 +40,6 @@ def relabel(adata, relabel_path, join_key="", sep=None):
     columns_to_drop = [col for col in adata.obs.columns if col.endswith('_y')]
     adata.obs.drop(columns=columns_to_drop, inplace=True)
     return adata
-
-def map_author_labels(obs, original_celltypes):
-    obs["new_dataset_title"] = obs["dataset_title"].apply(lambda x: x.replace(" ", "_")
-                                                                .replace("\\/", "_")
-                                                                .replace("(", "")
-                                                                .replace(")", "")
-                                                                .replace("\\", "")
-                                                                .replace("'", "")
-                                                                .replace(":", "")
-                                                                .replace(";", "")
-                                                                .replace("&", ""))
-
-    obs["new_observation_joinid"] = obs["new_dataset_title"].astype(str) + "_" + obs["observation_joinid"].astype(str)
-    
-    mapping = dict(zip(original_celltypes["new_observation_joinid"], original_celltypes["author_cell_type"]))
-    obs["author_cell_type"] = obs["new_observation_joinid"].map(mapping)
-
-    return obs
-
-def extract_data(cellxgene_obs_filtered, filtered_ids, subsample=10, organism=None, census=None, 
-    obs_filter=None, cell_columns=None, dataset_info=None, rename_file=None,
-    original_celltypes=None, seed=42):
-     
-    brain_cell_subsampled_ids = subsample_cells(cellxgene_obs_filtered, filtered_ids, subsample, seed=seed, organism=organism, rename_file=rename_file)
-    # Assuming get_seurat is defined to return an AnnData object
-    adata = cellxgene_census.get_anndata(
-        census=census,
-        organism=organism,
-        obs_value_filter=obs_filter,  # Ensure this is constructed correctly
-        obs_column_names=cell_columns,
-        obs_coords=brain_cell_subsampled_ids,
-        var_value_filter = "nnz > 10",
-        obs_embeddings=["scvi"])
-    
-    sc.pp.filter_genes(adata, min_cells=3)
-    sc.pp.filter_genes(adata, min_counts=200)
-    
-    print("Subsampling successful.")
-    newmeta = adata.obs.merge(dataset_info, on="dataset_id", suffixes=(None,"y"))
-    adata.obs = newmeta
-    
-    if isinstance(original_celltypes, pd.DataFrame) and not original_celltypes.empty:
-        adata.obs = map_author_labels(adata.obs, original_celltypes)
-    # Assuming relabel_wrapper is defined
-    # Convert all columns in adata.obs to factors (categorical type in pandas)
-    return adata
-
-
-def get_cellxgene_obs(census, organism, organ="brain", primary_data=True, disease="normal"):
-    value_filter = (
-        f"tissue_general == '{organ}' and "
-        f"is_primary_data == {str(primary_data)} and "
-        f"disease == '{disease}'"
-    )
-    return cellxgene_census.get_obs(census, organism, value_filter=value_filter)
-
-
-def get_census(census_version="2024-07-01", organism="homo_sapiens", subsample=5, assay=None, tissue=None, organ="brain",
-               ref_collections=["Transcriptomic cytoarchitecture reveals principles of human neocortex organization"," SEA-AD: Seattle Alzheimer’s Disease Brain Cell Atlas"], 
-               original_celltypes = None,
-               rename_file="/space/grp/rschwartz/rschwartz/cell_annotation_cortex.nf/meta/rename_cells.tsv",seed=42):
-
-    census = cellxgene_census.open_soma(census_version=census_version)
-    dataset_info = census.get("census_info").get("datasets").read().concat().to_pandas()
-    
-    cellxgene_obs = cellxgene_obs = get_cellxgene_obs(census, organism, organ=organ, primary_data=True, disease="normal")
-
-    cellxgene_obs = cellxgene_obs.merge(dataset_info, on="dataset_id", suffixes=(None,"_y"))
-    cellxgene_obs.drop(columns=['soma_joinid_y'], inplace=True)
-    cellxgene_obs_filtered = cellxgene_obs[cellxgene_obs['collection_name'].isin(ref_collections)]
-     
-    if assay:
-        cellxgene_obs_filtered = cellxgene_obs_filtered[cellxgene_obs_filtered["assay"].isin(assay)]
-    if tissue:
-        cellxgene_obs_filtered = cellxgene_obs_filtered[cellxgene_obs_filtered["tissue"].isin(tissue)]
-         
-    # Adjust organism naming for compatibility
-    organism_name_mapping = {
-        "homo_sapiens": "Homo sapiens",
-        "mus_musculus": "Mus musculus"
-    }
-    organism = organism_name_mapping.get(organism, organism)
-
-    cell_columns = [
-        "assay", "cell_type", "cell_type_ontology_term_id", "tissue",
-        "tissue_general", "suspension_type",
-        "disease", "dataset_id", "development_stage",
-        "soma_joinid", "observation_joinid"
-    ]
-    
-    if isinstance(original_celltypes, pd.DataFrame) and not original_celltypes.empty:
-       cellxgene_obs_filtered = map_author_labels(cellxgene_obs_filtered, original_celltypes)
-        
-    # Get embeddings for all data together
-    filtered_ids = cellxgene_obs_filtered['soma_joinid'].values
-    adata = extract_data(
-        cellxgene_obs_filtered, filtered_ids,
-        subsample=subsample, organism=organism,
-        census=census, obs_filter=None,
-        cell_columns=cell_columns, dataset_info=dataset_info, seed = seed,
-        original_celltypes=original_celltypes,
-        rename_file=rename_file
-    )
-    new_obs=rename_cells(adata.obs, rename_file=rename_file)
-    new_adata = adata[adata.obs["soma_joinid"].isin(new_obs["soma_joinid"])].copy()
-    new_obs = new_obs.set_index("soma_joinid")
-    new_adata.obs = new_adata.obs.set_index("soma_joinid")
-    new_adata.obs = new_adata.obs.loc[new_obs.index].copy()
-    new_adata.obs = new_obs
-     
-   # for name, ref in refs.items(): 
-    for col in new_adata.obs.columns:
-        if new_adata.obs[col].dtype.name =='category':
-            new_adata.obs[col] = pd.Categorical(new_adata.obs[col].cat.remove_unused_categories())
-    
-    return new_adata 
-
-
 
 def process_query(query, model_file_path, batch_key="sample", seed=42):
     scvi.settings.seed = seed
